+++ General/MediaInfo.vb
@@ -73,8 +73,8 @@ Public Class MediaInfo
                     at.Lossy = GetAudio(index, "Compression_Mode") = "Lossy"
                     at.SamplingRate = GetAudio(index, "SamplingRate").ToInt
                     at.BitDepth = GetAudio(index, "BitDepth").ToInt
-                    at.FormatString = GetAudio(index, "Format/String")
                     at.Format = GetAudio(index, "Format")
+                    at.FormatString = GetAudio(index, "Format/String")
                     at.FormatProfile = GetAudio(index, "Format_Profile")
                     at.Title = GetAudio(index, "Title").Trim
                     at.Forced = GetAudio(index, "Forced") = "Yes"